refactor: Logo filename to lowercase
